requirements.txt
telebot
spotipy
yt-dlp
requests
python-dotenv